realstate all property of storing and displaying in api

+++ database/migrations/2021_09_12_153159_create_locations_table.php
@@ -22,8 +22,9 @@ class CreateLocationsTable extends Migration
             $table->string('zipcode')->nullable();
             $table->string('city');
             $table->string('state');
-            $table->float('longitude');
-            $table->float('latitude');
+            $table->decimal('longitude',8,6);
+            $table->decimal('latitude',8,6);
+            // $table->decimal('foo', 5, 4);
             $table->bigInteger('country_id');
             $table->timestamps();
         });
